requirements.txt
Django==4.1.10
djangorestframework
django-cors-headers==4.1.0
django-mongodb-engine==0.6.0
djongo==1.3.6
pymongo==3.12.3